+++ buession-lang/src/main/java/com/buession/lang/Rawable.java
@@ -22,16 +22,22 @@
  * | Copyright @ 2013-2024 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
-package com.buession.core;
+package com.buession.lang;
 
 /**
  * 原始的，约定实现该接口的类，必须返回原始字节数组
  *
  * @author Yong.Teng
- * @since 2.0.0
+ * @since 3.0.0
  */
 @FunctionalInterface
-@Deprecated
-public interface Rawable extends com.buession.lang.Rawable {
+public interface Rawable {
+
+	/**
+	 * 返回原始的字节数组
+	 *
+	 * @return 原始的字节数组
+	 */
+	byte[] getRaw();
 
 }